refactor(evaluations): EvaluationCard.vue sous 300 lignes (H2)
Éclatement complet (pattern G1), API publique inchangée (props evaluation/syncing
+ emits create/edit/view-results/publish/preview/sync/delete, consommée par
TeacherEvaluations.vue) :
- EvaluationCardHeader.vue : titre + badges statut/online (CSS verbatim)
- EvaluationCardInfoGrid.vue : matière/classe/date/coef (CSS verbatim)
- EvaluationCardStatus.vue : fenêtre + bloc version en ligne (CSS verbatim)
- EvaluationCardActions.vue : boutons d'action + emits (CSS verbatim)
- EvaluationCard.vue : orchestrateur 642→53 lignes
Tests : EvaluationCard.test.js (existant, 3) + EvaluationCardSections (7) verts.
Build OK. Dette pré-existante #23 (formatDate local) conservée à l'identique.
